+++ python/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.4.0] - 2020-07-13
+
 ### Added
 
 - Option to disable processing of "style" tags. [#45](https://github.com/Stranger6667/css-inline/issues/45)
@@ -10,7 +12,6 @@
 ### Changed
 
 - Switch from `openssl` to `rustls` in `attohttpc` dependency. [#49](https://github.com/Stranger6667/css-inline/issues/49)
-
 
 ### Performance
 
@@ -63,7 +64,8 @@
 
 - Initial public release
 
-[Unreleased]: https://github.com/Stranger6667/css-inline/compare/python-v0.3.2...HEAD
+[Unreleased]: https://github.com/Stranger6667/css-inline/compare/python-v0.4.0...HEAD
+[0.4.0]: https://github.com/Stranger6667/css-inline/compare/python-v0.3.2...python-v0.4.0
 [0.3.2]: https://github.com/Stranger6667/css-inline/compare/python-v0.3.1...python-v0.3.2
 [0.3.1]: https://github.com/Stranger6667/css-inline/compare/python-v0.3.0...python-v0.3.1
 [0.3.0]: https://github.com/Stranger6667/css-inline/compare/python-v0.2.0...python-v0.3.0